13: Rename  and add comments explaining what all ES funcs need to do
